backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.1
requests==2.32.3
sqlalchemy
psycopg2-binary
python-dotenv==1.0.1
gunicorn==21.2.0
json-repair==0.8.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
